// web/src/api/generated/FeedRef.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Hash32 } from "./Hash32";
import type { NodeId } from "./NodeId";

/**
 * Entity referenced by a feed item.
 */
export type FeedRef = { "kind": "node", id: NodeId, } | { "kind": "app", name: string, } | { "kind": "block", height: number, } | { "kind": "tx", txid: Hash32, } | { "kind": "address", address: string, };